pkg: Move `kepval` packages to `legacy`
