+++ pkg/legacy/util/helpers.go
